pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-identity-engine"
version = "0.1.0"
description = "Synthetic ETL identity resolution prototype."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "ETL Identity Data Ruleset Engine Contributors" }
]
dependencies = [
  "pyarrow>=16.0.0"
]

[project.optional-dependencies]
dev = [
  "PyYAML>=6.0.2",
  "pytest>=8.0.0",
  "ruff>=0.5.0"
]

[project.scripts]
etl-identity-engine = "etl_identity_engine.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
